Print out UID of pod during vn2 deployment
When things go wrong, we may not get a container ID printout, but the container
group is always the pod UID plus "_0".  By printing the pod UID, we makes it
easier to diagnose issues after the fact.

This UID is available as soon as the pod creation request is accepted by k8s:

    Waiting for pods to be in Running state...
    Running command: kubectl get pods -l app=test -o json
    Pod test-bf6545c76-9fvh5 status: Pending
    Uid: da2d4760-78b5-4c46-9fe2-36e9c8fcd2f5
    Conditions: PodReadyToStartContainers, Initialized, Ready, ContainersReady, PodScheduled

Using this, one can simply search in the Azure portal for the in-progress
container group, or search for why a container creation failed in the vn2 logs.